Add advisory_lock utility from AWX (#661)
This migrates `advisory_lock` from AWX to DAB. This is similar in spirit
to https://github.com/ansible/django-ansible-base/pull/660

Importantly, I went the route of _not_ adding a new dependency, because
there was only a single method from the library we were using, and I'm
pretty confident that the MIT license lets you do whatever you want. In
doing this, I deleted the use of `six` for python2 compatibility which
further reduced the complexity of the code I ported.