AAP-39645 Allow non-admin admin (#716)
Fixes issues blocking a user not named "admin" from being the admin
user. The issue in this case was creation of the db authenticator via
"aap-gateway-manage authenticators --initialize" was hard coded to the
user "admin" but this PR allows it to be created by the system user or
by no user if there is no "admin" user or system user defined.

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update
- [ ] Test update
- [ ] Refactoring (no functional changes)
- [ ] Development environment change
- [ ] Configuration change

- [x] I have performed a self-review of my code
- [x] I have added relevant comments to complex code sections
- [ ] I have updated documentation where needed
- [ ] I have considered the security impact of these changes
- [ ] I have considered performance implications
- [ ] I have thought about error handling and edge cases
- [x] I have tested the changes in my local environment

1. Start gateway (with PR for AAP-39645 + this PR) after modifying
gateway_admin_username in container-startup.yml to something other than
admin.
2. See that everything works OK (there were errors in the log and didn't
start before). Log in as the specified user and see they have system
administrator role.
3. It's not really possible to unset SYSTEM_USER for gateway or for
test_app, both are not happy without one. This situation could be tested
with awx (since apparently it does not use a system user) but that is a
challenge to set up. Unit tests cover this scenario.

See above.

N/A

- [ ] Requires documentation updates
- [ ] Requires downstream repository changes
- [ ] Requires infrastructure/deployment changes
- [ ] Requires coordination with other teams
- [ ] Blocked by PR/MR: #XXX

N/A

---------

Co-authored-by: Bryan Havenstein <bhavenst@redhat.com>