AAP-43949 Add psycopg db utils for dispatcherd use (#715)
## Description
We have pushed out our objective merge date for DAB task in the general,
which is at https://github.com/ansible/django-ansible-base/pull/702, and
doing this creates a need to be able to feature-flag a release using
https://github.com/ansible/dispatcherd

Using dispatcherd by itself doesn't need most of the content from DAB
task, but it needs _some utility_ type methods. As we flip-flopped on
the exact order of operations for this, I had previously put up
https://github.com/ansible/django-ansible-base/pull/686, and closed it
_in favor of_ DAB task. Some content was added while I worked on DAB
task, so I created a new branch (this one) and put the updated content
there.

Why?

Merging this will allow the EDA-server and AWX work to de-duplicate code
that was copy+pasted there.


https://github.com/ansible/awx/compare/devel...feature_dispatcher#diff-60fa214144754e9e3cf12cc5c7054d8b0ce04f43da33181c0fe9f0d927627038


https://github.com/ansible/eda-server/compare/main...dispatcher-poc-v2#diff-ff14e1e048588b54cbac15edc61d4c830abc8a87cdae26b7fb603d52f3fe1960

Those were always intended to go live in DAB.

Next question - why can't we just put them in the dispatcherd repo?
Because _it does not have Django_. This is very important. I do intend
to write some docs in dispatcherd that will reference this content.

Ping @bzwei for review

## Type of Change
- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update
- [x] Test update
- [x] Refactoring (no functional changes)
- [ ] Development environment change
- [ ] Configuration change

## Self-Review Checklist
- [x] I have performed a self-review of my code
- [x] I have added relevant comments to complex code sections
- [ ] I have updated documentation where needed
- [ ] I have considered the security impact of these changes
- [ ] I have considered performance implications
- [ ] I have thought about error handling and edge cases
- [x] I have tested the changes in my local environment

## Testing Instructions
This probably needs to be tested with the respective EDA or AWX
branches. They have been working, this just consolidates these methods.

### Prerequisites
<!-- List any specific setup required -->

### Steps to Test
1. 
2. 
3. 

### Expected Results
<!-- Describe what should happen after following the steps -->

## Additional Context
<!-- Optional but helpful information -->

### Required Actions
- [ ] Requires documentation updates
- [x] Requires downstream repository changes
- [ ] Requires infrastructure/deployment changes
- [x] Requires coordination with other teams
- [ ] Blocked by PR/MR: #XXX

### Screenshots/Logs
<!-- Add if relevant to demonstrate the changes -->