AAP-36381: Make SAML username/permid required (#719)
## Description
Some time back, attr_user_permanent_id and attr_username were made
nullable/not-required since the docs make it seem like that is
acceptable. It however is not and authentication fails without either of
them. The docs have already been updated to state they are required, but
the UI still shows them as optional. This just makes them required in
the UI.


## Type of Change
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update
- [ ] Test update
- [ ] Refactoring (no functional changes)
- [ ] Development environment change
- [ ] Configuration change

## Self-Review Checklist
- [x] I have performed a self-review of my code
- [ ] I have added relevant comments to complex code sections
- [ ] I have updated documentation where needed
- [ ] I have considered the security impact of these changes
- [ ] I have considered performance implications
- [ ] I have thought about error handling and edge cases
- [ ] I have tested the changes in my local environment

## Testing Instructions
### Prerequisites

### Steps to Test
1. Try to create a SAML authenticator without specifying one or both of
these params.
2. See that it's not accepted.
3. 

### Expected Results
Impossible to create SAML authenticator without username and permanent
ID fields.

## Additional Context
N/A

---------

Co-authored-by: Bryan Havenstein <bhavenst@redhat.com>